obtain the ranking of each Pokémon

+++ src/main/java/tripleusage/com/usage/domain/MoveTotal.java
@@ -2,5 +2,5 @@ package tripleusage.com.usage.domain;
 
 public interface MoveTotal {
     String getMove();
-    int getMovetotal();
+    float getMovetotal();
 }
